Handle unrecognised roles in api's user controller
